+++ src/Form/ReplyType.php
@@ -1,0 +1,41 @@
+<?php
+
+namespace App\Form;
+
+use App\Entity\Reply;
+use Symfony\Component\Form\AbstractType;
+use Symfony\Component\Form\Extension\Core\Type\FileType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
+use Symfony\Component\Form\Extension\Core\Type\TextareaType;
+use Symfony\Component\Form\Extension\Core\Type\TextType;
+use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\OptionsResolver\OptionsResolver;
+
+class ReplyType extends AbstractType
+{
+    public function buildForm(FormBuilderInterface $builder, array $options): void
+    {
+        $builder
+            ->add('content', TextareaType::class, [
+                'attr' => [
+                    'placeholder' => 'Votre réponse'
+                ]
+            ])
+            ->add('image', FileType::class, [
+                'label' => 'Image de la réponse',
+                'mapped' => false,
+                'required' => false
+            ])
+            ->add('submit', SubmitType::class, [
+                'label' => 'Répondre'
+            ])
+        ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $resolver->setDefaults([
+            'data_class' => Reply::class,
+        ]);
+    }
+}